Move component OAuth test case
